+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bugs fixed
 
     - Fix taps "search value on flows"
+    - Fix middleware not working with nrepl > 1.3.0-beta2
 
 ## 4.0.2 (25-11-2024)